refactor(extension): split src/mergify.js into ES modules with esbuild bundle
Splits the monolithic 2500-line src/mergify.js into focused modules
(logo, debug, dom, cache, timestamps, queue, stacks) and adds an
esbuild step that bundles them back into the single
dev-build/mergify.js the Manifest V3 content script declares. No
behavior change. Tests use Jest with @swc/jest to handle the ESM
imports.

src/mergify.js becomes a small orchestrator that wires the
MutationObserver, the URL-change reset, and the tryInject loop, and
re-exports everything for the test suite (which keeps importing from
"../mergify").
