+++ jest.config.js
@@ -2,4 +2,7 @@ module.exports = {
     testEnvironment: "jsdom",
     roots: ["<rootDir>/src"],
     testMatch: ["**/*.test.js"],
+    transform: {
+        "^.+\\.js$": ["@swc/jest"],
+    },
 };
