+++ package.json
@@ -12,6 +12,9 @@
     },
     "devDependencies": {
         "@biomejs/biome": "2.4.14",
+        "@swc/core": "^1.15.32",
+        "@swc/jest": "^0.2.39",
+        "esbuild": "^0.28.0",
         "jest": "^30.3.0",
         "jest-environment-jsdom": "^30.3.0",
         "nodemon": "^3.1.14"
