+++ src/debug.js
@@ -1,0 +1,6 @@
+window.__MERGIFY_DEBUG__ ??= false;
+
+export function debug(...args) {
+    if (!window.__MERGIFY_DEBUG__) return;
+    console.log(...args);
+}
