+++ src/dom.js
@@ -1,0 +1,77 @@
+// Matches timestamps generated by the Mergify engine in merge queue reports.
+export const MERGIFY_TIMESTAMP_REGEX =
+    /^(\d{4}-\d{2}-\d{2}) (\d{2}:\d{2}) UTC$/;
+export const TIMESTAMP_CONVERTED_ATTR = "data-mergify-local-time";
+
+const LOCAL_TIME_FORMATTER = new Intl.DateTimeFormat(undefined, {
+    year: "numeric",
+    month: "short",
+    day: "numeric",
+    hour: "2-digit",
+    minute: "2-digit",
+    timeZoneName: "short",
+});
+
+export function formatLocalTime(date) {
+    return LOCAL_TIME_FORMATTER.format(date);
+}
+
+export function isMergifyBotComment(container) {
+    const authorLink = container.querySelector(
+        'a.author[href="/apps/mergify"]',
+    );
+    if (authorLink) return true;
+
+    const avatar = container.querySelector("img.avatar");
+    if (avatar && /mergify/i.test(avatar.getAttribute("src") || "")) {
+        return true;
+    }
+
+    return false;
+}
+
+export function getPullRequestData() {
+    const url = new URL(window.location.href);
+    const parts = url.pathname.split("/");
+    return {
+        org: parts[1],
+        repo: parts[2],
+        pull: parts[4],
+        subpath: parts[5] || "",
+    };
+}
+
+export function isGitHubPullRequestPage() {
+    const url = new URL(window.location.href);
+    const parts = url.pathname.split("/");
+    return parts.length >= 5 && parts[3] === "pull";
+}
+
+export function findFirstElement(selectors) {
+    for (const sel of selectors) {
+        const el = document.querySelector(sel);
+        if (el) return el;
+    }
+    return null;
+}
+
+export function findTimelineActions() {
+    const mergeBoxDiv = document.querySelector(
+        "div[class=discussion-timeline-actions]",
+    );
+    return mergeBoxDiv;
+}
+
+export const PR_STATUS_SELECTORS = [
+    { selector: 'span[data-status="draft"]', status: "draft" },
+    { selector: 'span[data-status="pullOpened"]', status: "open" },
+    { selector: 'span[data-status="pullMerged"]', status: "merged" },
+    { selector: 'span[data-status="pullClosed"]', status: "closed" },
+];
+
+export function readPrStatusFromDocument(scope) {
+    for (const { selector, status } of PR_STATUS_SELECTORS) {
+        if (scope.querySelector(selector)) return status;
+    }
+    return null;
+}
